requirements.txt
# Core ML/NLP Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0
accelerate>=0.20.0

# Data Science & Visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
scikit-learn>=1.3.0

# Web Interface
streamlit>=1.25.0
streamlit-option-menu>=0.3.0

# Evaluation Metrics
sacrebleu>=2.3.0
rouge-score>=0.1.2
bert-score>=0.3.13
unbabel-comet>=2.0.0

# Advanced Features
bitsandbytes>=0.41.0
optimum>=1.12.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
pathlib2>=2.3.7
typing-extensions>=4.7.0
requests>=2.31.0
urllib3>=2.0.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional GPU support
# torch-audio>=2.0.0  # Uncomment for audio processing
# torch-vision>=0.15.0  # Uncomment for vision tasks
